Updated implementation of `ast::MetaIntrinsic`
- Previously the ID of the intrinsic was part of the standard
  library. The implementation has been updated to use the choice type
  in the `meta` module

+++ src/ast/meta/intrinsic.hpp
@@ -6,7 +6,7 @@
 namespace qat::ast {
 
 // NOTE - !!! Update the standard library if this is updated
-enum class IntrinsicID : u32 {
+enum class IntrinsicID : u8 {
 	matrix_multiply = 0,
 	matrix_transpose,
 	matrix_column_major_load,
